word2vec/requirements.txt
gensim
Flask
mecab-python3